# server/controllers/SchoolMS_db/generated/LessonControllerGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE LessonControllerGenerated.js PLEASE EDIT ../LessonController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import LessonModel from "../../../models/SchoolMS_db/LessonModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import LessonController from "../LessonController";

const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/lesson`;
    router.post(baseUrl + "", authorize([]), LessonController.create);
    router.delete(baseUrl + "/:id", authorize([]), LessonController.delete);
    router.get(baseUrl + "/findBy_Subject/:key", authorize([]), LessonController.findBy_Subject);
    router.get(baseUrl + "/findBy_Teacher/:key", authorize([]), LessonController.findBy_Teacher);
    router.get(baseUrl + "/findByassignment/:key", authorize([]), LessonController.findByassignment);
    router.get(baseUrl + "/findByclass/:key", authorize([]), LessonController.findByclass);
    router.get(baseUrl + "/findBydone/:key", authorize([]), LessonController.findBydone);
    router.get(baseUrl + "/findBylessonname/:key", authorize([]), LessonController.findBylessonname);
    router.get(baseUrl + "/findBylessonno/:key", authorize([]), LessonController.findBylessonno);
    router.get(baseUrl + "/findBysubjectname/:key", authorize([]), LessonController.findBysubjectname);
    router.get(baseUrl + "/findByteachername/:key", authorize([]), LessonController.findByteachername);
    router.get(baseUrl + "/findBytest/:key", authorize([]), LessonController.findBytest);
    router.get(baseUrl + "/:id", authorize([]), LessonController.get);
    router.get(baseUrl + "/:id/get_Subject", authorize([]), LessonController.get_Subject);
    router.get(baseUrl + "/:id/get_Subject", authorize([]), LessonController.get_Subject);
    router.get(baseUrl + "/:id/get_Teacher", authorize([]), LessonController.get_Teacher);
    router.get(baseUrl + "", authorize([]), LessonController.list);
    router.post(baseUrl + "/:id", authorize([]), LessonController.update);
  },


  // CRUD METHODS


  /**
  * LessonModel.create
  *   @description CRUD ACTION create
  *
  */
  create: async (req, res) => {
    try {
      const result = await LessonModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LessonModel.delete
  *   @description CRUD ACTION delete
  *   @param String id id
  *   @param String lessonname
  *
  */
  delete: async (req, res) => {
    try {
      const result = await LessonModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LessonModel.findBy_Subject
  *   @description CRUD ACTION findBy_Subject
  *
  */
  findBy_Subject: async (req, res) => {
    try {
      const result = await LessonModel.findBy_Subject(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LessonModel.findBy_Teacher
  *   @description CRUD ACTION findBy_Teacher
  *
  */
  findBy_Teacher: async (req, res) => {
    try {
      const result = await LessonModel.findBy_Teacher(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LessonModel.findByassignment
  *   @description CRUD ACTION findByassignment
  *
  */
  findByassignment: async (req, res) => {
    try {
      const result = await LessonModel.findByassignment(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LessonModel.findByclass
  *   @description CRUD ACTION findByclass
  *
  */
  findByclass: async (req, res) => {
    try {
      const result = await LessonModel.findByclass(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LessonModel.findBydone
  *   @description CRUD ACTION findBydone
  *
  */
  findBydone: async (req, res) => {
    try {
      const result = await LessonModel.findBydone(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LessonModel.findBylessonname
  *   @description CRUD ACTION findBylessonname
  *
  */
  findBylessonname: async (req, res) => {
    try {
      const result = await LessonModel.findBylessonname(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LessonModel.findBylessonno
  *   @description CRUD ACTION findBylessonno
  *
  */
  findBylessonno: async (req, res) => {
    try {
      const result = await LessonModel.findBylessonno(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LessonModel.findBysubjectname
  *   @description CRUD ACTION findBysubjectname
  *
  */
  findBysubjectname: async (req, res) => {
    try {
      const result = await LessonModel.findBysubjectname(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LessonModel.findByteachername
  *   @description CRUD ACTION findByteachername
  *
  */
  findByteachername: async (req, res) => {
    try {
      const result = await LessonModel.findByteachername(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LessonModel.findBytest
  *   @description CRUD ACTION findBytest
  *
  */
  findBytest: async (req, res) => {
    try {
      const result = await LessonModel.findBytest(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LessonModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id resource
  *   @param String lessonname
  *
  */
  get: async (req, res) => {
    try {
      const result = await LessonModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LessonModel.get_Subject
  *   @description CRUD ACTION get_Subject
  *
  */
  get_Subject: async (req, res) => {
    try {
      const result = await LessonModel.get_Subject(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LessonModel.get_Subject
  *   @description CRUD ACTION get_Subject
  *   @param String subjectname
  *
  */
  get_Subject: async (req, res) => {
    try {
      const result = await LessonModel.get_Subject(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LessonModel.get_Teacher
  *   @description CRUD ACTION get_Teacher
  *   @param String teachername
  *
  */
  get_Teacher: async (req, res) => {
    try {
      const result = await LessonModel.get_Teacher(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * LessonModel.list
  *   @description CRUD ACTION list
  *
  */
  list: async (req, res) => {
    try {
      const result = await LessonModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * LessonModel.update
  *   @description CRUD ACTION update
  *   @param String id id
  *   @param String lessonname
  *
  */
  update: async (req, res) => {
    try {
      const result = await LessonModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;