# server/swagger.yaml
openapi: 3.0.0
info:
  title: SchoolMS
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /lesson/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Lesson
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: string
        - name: lessonname
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Lesson
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
        - name: lessonname
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    post:
      summary: CRUD ACTION update
      tags:
        - Lesson
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: string
        - name: lessonname
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Lesson"
...
  /lesson/{id}/get_Subject:
    get:
      summary: CRUD ACTION get_Subject
      tags:
        - Lesson
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get_Subject
      tags:
        - Lesson
      parameters:
        - name: subjectname
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /lesson/{id}/get_Teacher:
    get:
      summary: CRUD ACTION get_Teacher
      tags:
        - Lesson
      parameters:
        - name: teachername
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    
  #-----  Student  -----
...
  /student/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Student
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: string
        - name: studentname
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Student
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: string
        - name: studentname
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    post:
      summary: CRUD ACTION update
      tags:
        - Student
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: string
        - name: studentname
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Student"
...
  /student/{id}/get_Class:
    get:
      summary: CRUD ACTION get_Class
      tags:
        - Student
      parameters:
        - name: class
          in: path
          description: 
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
  /student/{id}/get_Exam:
    get:
      summary: CRUD ACTION get_Exam
      tags:
        - Student
      parameters:
        - name: examname
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    
  #-----  Subject  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Lesson:
      properties:
        _id:
          type: string
        assignment:
          type: boolean
        class:
          type: number
        done:
          type: boolean
        lessonname:
          type: string
        lessonno:
          type: number
        subjectname:
          type: string
        teachername:
          type: string
        test:
          type: boolean
      required:
        -  _id
        - class
        - lessonno
        - subjectname
        - teachername
    Student:
      properties:
        _id:
          type: string
        class:
          type: number
        dob:
          type: integer
        fathersname:
          type: string
        mothersname:
          type: string
        percentinlastexam:
          type: number
        rollno:
          type: string
        studentname:
          type: string
      required:
        -  _id
        - class
        - rollno
        - studentname
...